feat: DTO 생성
Added: BasicGameService and DTO
Added: BasicGameCategoryService and DTO
Added: Edit unit of price on every table

+++ src/main/java/com/teemo/shopping/Order/domain/CardPayment.java
@@ -20,7 +20,7 @@ import lombok.NoArgsConstructor;
 @DiscriminatorValue(Values.KAKAOPAY)
 public class CardPayment extends Payment {
     @Builder
-    public CardPayment(double price, PaymentStatus status) {
+    public CardPayment(int price, PaymentStatus status) {
         super(price, status);
     }
 }
